refactor: manage custom Express server with kiliman's packages
