+++ app/routes/[_].healthcheck.ts
@@ -1,7 +1,8 @@
 import type { LoaderFunctionArgs } from '@remix-run/node';
+import { invariant } from '@epic-web/invariant';
 
 export async function loader({ request }: LoaderFunctionArgs) {
-  const host = request.headers.get('X-Forwarded-Host') ?? request.headers.get('host');
+  const host = request.headers.get('X-Forwarded-Host') || request.headers.get('host');
   const url = new URL('/', `http://${host}`);
 
   try {
@@ -9,7 +10,7 @@ export async function loader({ request }: LoaderFunctionArgs) {
       method: 'HEAD',
       headers: { 'x-from-healthcheck': 'true' },
     });
-    if (!response.ok) throw new Error(`HEAD request failed to ${url}`);
+    invariant(response.ok, `HEAD request failed to ${url}`);
     return new Response('OK', { status: 200 });
   } catch (error) {
     console.log('❌ Healthcheck failed', { error });
