+++ package.json
@@ -23,11 +23,11 @@
     "build:remix": "remix vite:build",
     "deploy": "flyctl deploy --remote-only",
     "dev": "run-s clean dev:remix",
-    "dev:remix": "tsx watch --clear-screen=false --inspect --require dotenv/config --watch-path ./server.ts ./server.ts",
+    "dev:remix": "vite",
     "format": "prettier --cache --write .",
     "format:check": "prettier --cache --check .",
     "lint": "eslint --cache --cache-location ./node_modules/.cache/eslint .",
-    "start": "cross-env NODE_ENV=production tsx ./server.ts",
+    "start": "cross-env NODE_ENV=production node ./build/server/index.js",
     "typecheck": "tsc",
     "validate": "run-p format:check lint typecheck"
   },
@@ -36,24 +36,20 @@
     "@fontsource/source-sans-3": "5.1.0",
     "@remix-pwa/sw": "3.0.10",
     "@remix-pwa/worker-runtime": "2.1.4",
-    "@remix-run/express": "2.13.1",
     "@remix-run/node": "2.13.1",
     "@remix-run/react": "2.13.1",
     "@remix-run/serve": "2.13.1",
     "algoliasearch": "5.12.0",
     "clsx": "2.1.1",
     "cross-env": "7.0.3",
-    "express": "4.21.1",
     "isbot": "5.1.17",
     "lodash-es": "4.17.21",
     "marked": "14.1.3",
-    "morgan": "1.10.0",
     "pretty-cache-header": "1.0.0",
     "react": "18.3.1",
     "react-dom": "18.3.1",
     "remix": "2.13.1",
-    "source-map-support": "0.5.21",
-    "tsx": "4.19.2",
+    "remix-create-express-app": "0.4.5",
     "zod": "3.23.8"
   },
   "devDependencies": {
@@ -77,6 +73,7 @@
     "postcss": "8.4.47",
     "prettier": "3.3.3",
     "prettier-plugin-tailwindcss": "0.6.8",
+    "remix-express-dev-server": "0.4.4",
     "semantic-release": "24.2.0",
     "tailwindcss": "3.4.14",
     "typescript": "5.6.3",
